requirements.txt

### File: requirements.txt

```text
requests
six
fake_useragent
pysocks
